Move xml helper functions to the TokenPair object so it can be reused
